feat: Add dynamic charts such as donut charts and bar charts, and reorganize folders to create live cards #75
