+++ badge_reader/config/config.exs
@@ -82,7 +82,3 @@ config :phoenix, :json_library, Jason
 # Import environment specific config. This must remain at the bottom
 # of this file so it overrides the configuration defined above.
 import_config "#{config_env()}.exs"
-
-config :live_charts,
-  adapter: LiveCharts.Adapter.ApexCharts,
-  json_library: Jason
